Test for command type, not architecture
This patch modifies the way available commands are tested; If a Mac user installs and uses GNU ls it is correctly detected and used.
the ls_with_file_shortcuts now works on ZSH. It was failing with "command not found" due to parameters passed in $_ll_command which were interpreted as a command name.