Fix for numbered directories or files
This commit fixes some odd behavior for numbered files and
directories. Even if the environment variables is undefined
cd'ing into a numbered directory is not possible. This pathc
fixes this by checking if the argument is a valid
file or directory and is numnbered and if this is the case will
only output the argument values instead of a substitution with $e..

In case there exist both, numbered valid argument and the
directory or folder, the file will always win to avoid clashes with
standard wrapped functions like cd, rm, mv...