Fix quoting of arguments with special characters
If you had a file with a character that had a special meaning in a shell
command line, scmb_expand_args would not requote the string before it
was run through eval.

Here is a an example problem:

touch 'foo;bar'
rm 'foo;bar'

This would try to run "rm foo;bar" and get an error that "foo" doesn't
exist and "bar" is an invalid command.