aliases: consistent indentation and quotation
this is unrelated to the previous commit and zero functionality but
while I was changing things I wanted to clean up the indentation and
use consistent single quotes versus double quotes in the functions
calls for `_alias()` and `_git_alias()`.

In my opinion this makes the file more readable, but if it bothers you
let me know and I’ll be happy to remove it from the pull request.